+++ MultiHotelManagement/src/main/java/com/MHM/MultiHotelManagement/dto/response/PaymentResponseDTO.java
@@ -1,6 +1,8 @@
 package com.MHM.MultiHotelManagement.dto.response;
 
 import lombok.Data;
+
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -8,7 +10,7 @@ public class PaymentResponseDTO {
 
     private Long id;
     private String method;
-    private Double amount;
+    private BigDecimal amount;
     private String status;
     private Long bookingId;
     private String bookingReference;
