+++ MultiHotelManagement/src/main/java/com/MHM/MultiHotelManagement/dto/response/WalletResponseDTO.java
@@ -2,6 +2,7 @@ package com.MHM.MultiHotelManagement.dto.response;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -9,8 +10,8 @@ public class WalletResponseDTO {
     private Long id;
     private Long userId;
     private String userName;
-    private Double balance;
-    private Double totalEarned;
-    private Double totalWithdrawn;
+    private BigDecimal balance;
+    private BigDecimal totalEarned;
+    private BigDecimal totalWithdrawn;
     private LocalDateTime createdAt;
 }
